+++ src/App.jsx
@@ -1,29 +1,21 @@
-import { useState } from 'react'
+
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
+import React from 'react';
 import './App.css'
+import TodoList from './TodoList'
 
 
-const todoList=[
-  { id:0 , title:"Study the  React Book"},
-  { id:1 , title:"Watch Lesson's Video"},
-  { id:2 , title:"Participate Mentor group meeting"},
-  { id:3 , title:"Submit the assignment"}];
+
 
 function App() {
  
   return (
     <>
-      <h1>Todo List</h1>
-      <ul >
-        {/* Map over the todoList array */}
-
-      {todoList.map(({id , title})=>
-      ( <li key={id}>{title}</li>)
-      )}
-      </ul>
+      <h1>Todo List </h1>
+      <TodoList /> 
     </>
-  )
+  );
 }
 
 export default App
